Refactor data types and update deserialization logic
Changed data types of Quantity, Calories, and Proteins from int to double in AddMealRequest and Meal classes. Updated JSON schema in AIService.cs accordingly. Introduced MealDto for deserialization and adjusted logic to use it. Reordered operations in ReportService.cs to update user before AI call.

+++ CompanionAPI/Contracts/MealContracts/AddMealRequest.cs
@@ -3,8 +3,8 @@
 public record AddMealRequest(
     string UserId,
     string Name,
-    int Quantity,
-    int Calories,
-    int Proteins,
+    double Quantity,
+    double Calories,
+    double Proteins,
     string Unit
 );
